Extract SimpleBackendConfig and SimpleBackendConfigBuilder
These base classes can be extended to more easily create configs for
new backends.
